+++ templates/events/live_participant.njk
@@ -297,7 +297,7 @@
 		});
 
 		socket.on('end-quiz', () => {
-			window.location.href = '/live-results';
+			window.location.href = '/live/results';
 		});
 
 		window.addEventListener('load', () => showMessage(waitingMsg));
